route added, snackbar theme file, custom widget segregate
