+++ lib/home_page/bloc/home_page_event.dart
@@ -1,0 +1,13 @@
+class Event {
+  final String title;
+  final String dateRange;
+  final String imagePath;
+  final String location;
+
+  Event({
+    required this.title,
+    required this.dateRange,
+    required this.imagePath,
+    required this.location,
+  });
+}
