+++ lib/home_page/bloc/home_page_state.dart
@@ -1,0 +1,36 @@
+import 'package:akzonobel/resources/images.dart';
+
+import 'home_page_event.dart';
+
+final List<Event> events = [
+  Event(
+    title: 'Testing event',
+    dateRange: '19 Jul | 23 Jul - 2024',
+    imagePath: Images.banner,
+    location: 'Ahmedabad',
+  ),
+  Event(
+    title: 'Test event',
+    dateRange: '15 Feb | 18 Feb - 2025',
+    imagePath: Images.banner_2,
+    location: 'Intelivita,Ahmedabad',
+  ),
+  Event(
+    title: 'Test event',
+    dateRange: '15 Feb | 18 Feb - 2025',
+    imagePath: Images.banner_3,
+    location: 'Intelivita,Ahmedabad',
+  ),
+  Event(
+    title: 'Testing event',
+    dateRange: '19 Jul | 23 Jul - 2024',
+    imagePath: Images.banner_4,
+    location: 'Ahmedabad',
+  ),
+  Event(
+    title: 'Testing event',
+    dateRange: '19 Jul | 23 Jul - 2024',
+    imagePath: Images.banner_5,
+    location: 'Ahmedabad',
+  ),
+];
