+++ lib/utils/app_utils.dart
@@ -5,8 +5,8 @@ import 'package:flutter/material.dart';
 class AppUtils{
   AppUtils._();
 
-  static final AppUtils _instance = AppUtils._();
-  static AppUtils get instance => _instance;
+  static final AppUtils _instance = AppUtils._(); // this is singleton(Only one instance of a class is created in the entire application,) instance
+  static AppUtils get instance => _instance;  // this is used in whole project where we need to use this AppUtils(public way to access it)
 
 
   void showSnackBar(
