+++ pubspec.yaml
@@ -35,6 +35,7 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
   flutter_bloc: ^9.1.0
+  chopper: ^8.1.0
 
 dev_dependencies:
   flutter_test:
